feat(services/intl): add `primaryLocale` alias (#726)
This PR add an alias to `locale.firstObject` to the `intl` service called `primaryLocale`. This alias improves code readability when accessing the primary locale and eases the transition from `ember-i18n` to `ember-intl`: `i18n.locale` -> `intl.primaryLocale`

Closes https://github.com/ember-intl/ember-intl/issues/695 for good!

I've added documentation regarding this new property and updated the `locale` documentation, as it's always an array.