Add support for Ember SafeString
Without this fix the output in the test example would be

```
<strong>Welcome ,<em>Alexander</em>,!</strong>
```

Probably because the AST is an array and if a member of the AST isn't a string it will cast the AST into a string, instead of concatenating everything.

Not sure if this is the ideal place to address this, but seemed like a reasonable approach.